rtc: virtual hardware model of a real-time clock.
This is a standalone virtual hardware model of Motorola's MC146818
real-time clock. The device can be controlled through CMOS I/O. This
is illustrated in the accompanying tests.

Note that this model has its roots in QEMU. In particular, most of
the source code is by Fabrice Bellard and Anthony Liguori.

This standalone model makes several simplifying assumptions. For
example, it is not attached to an ISA bus to avoid building a tree
of QEMU devices. In particular, QEMU's i440fx PCI host is not needed.
In addition, no i8259 interrupt controller is required. However, as
the result, programmable periodic interrupts cannot be supported.