ethoc: Do not raise IRQ when INT_MASK is changed
We only seek to analyse features that are justified by the data sheet. In the
OpenCores Ethernet MAC data sheet, the possible reasons for interrupts are
defined in Section 3.2 (p. 10). None of these would suggest that an interrupt
can be raised when the INT_MASK register is changed. We conclude that this is
QEMU-specific functionality which we exclude from our analysis with CBMC.