Add experimental results for different partial-order encodings and SAT solvers
We performed all experiments on a 64-bit machine running Linux 3.17.4 with eight
Intel Xeon 3.07 GHz cores and 48 GB of main memory. We do not impose any memory
quota which means the entire amount of memory is available to CBMC. We used a
modified version of CBMC 5.2, released on 29 August 2015. The required patches
for these modifications can be found in the "cbmc-5.2-patches" directory.

We experimentally compare the cubic-size and quadratic-size partial-order encoding.
For the SAT solvers, we use MiniSAT 2.0 and Glucose 4.0 (also known as Glucose-Syrup).

Note that we trimmed down the code of the HW/SW benchmark because the analysis runs
out of memory anyways. The required patch can be found in the "ethoc-hw-sw-patches"
directory.