Add experimental results for HW.1-5 benchmark where we measure lcnf
We have annotated the satcheck_glucose_baset<T>::lcnf(const bvt &)
function in src/solvers/sat/satcheck_glucose.cpp such that it prints
out the number of times it has been called and the system time in
seconds. This shows that only 1389 (= 1441935975 - 1441934586)
seconds (< 30 minutes) are spent adding clauses to Glucose.